websockets/requirements.txt
djangorestframework==3.15.2
Django==4.1.7
channels_redis
channels  
python-decouple
mysqlclient